Moved example problem to new file
